13-DeploymentFinaldep/requirements.txt
alembic==1.13.2
attrs==23.1.0
autopep8==2.0.1
Babel==2.15.0
bcrypt==4.0.1  # Downgraded to avoid compilation issues
bidict==0.22.1
blinker==1.8.2
certifi>=2024.8.30
  # Updated to avoid compilation error
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cowsay==6.0
cryptography==36.0.0  # Updated to avoid compilation issues
cycler==0.12.1
distlib==0.3.8
distro==1.7.0
dnspython==2.4.2
email_validator==2.2.0
eventlet==0.33.3
filelock==3.16.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.8
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.51.0
gevent==23.9.0.post1
greenlet==3.0.3
gunicorn==20.1.0
httplib2==0.20.2
idna==2.6
importlib-metadata==6.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
Mako==1.3.5
Markdown==3.3.6
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.8.4
more-itertools==8.10.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.0
packaging==24.0
pillow==10.3.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pycparser==2.21
pygame==2.5.2
Pygments==2.11.2
PyJWT==2.3.0
pymongo==4.5.0
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.7.0
python-socketio==5.9.0
pytz==2024.1
  # Updated to avoid build errors
requests>=2.18.4
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.32
supervisor==4.2.1
systemd-python==234
tomli==2.0.1
typing_extensions==4.12.2
urllib3>=1.22
virtualenv==20.26.6
wadllib==1.3.6
Werkzeug==3.0.3
whitenoise==6.7.0
WTForms==3.1.2
zipp==3.12.1
zope.event==5.0
zope.interface==6.0
setuptools>=58.0.0
wheel>=0.36.0

